moved more out of main
